Minor Patch.
ApplicationHelpers:
Moved SetThreadAffinity here, to make it reusable.

CpuGremlins:
This version is now called CpuUsageGremlins. Keeping the gremlins focused on one specific activity.

ThreadContainer:
Added the ability to asign Custom AsyncFunc workloads.
Added the Throttling to lower CPU Usage.
And the ability to Terminate a thread that is Processing gracefully.